ref #16 used lambda function in Show Post fragment and view model and also Home button in Show post fragment
