+++ app/src/main/java/com/example/project_1/ui/addUser/AddUserViewModel.kt
@@ -11,5 +11,4 @@ class AddUserViewModel(application : Application): AndroidViewModel(application)
     fun addUser( user : UserDetailData){
         userDao.insertUserDetail(user)
     }
-
 }
