+++ app/src/main/java/com/example/project_1/ui/showPost/ShowPostFragment.kt
@@ -5,12 +5,16 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Button
 import android.widget.Toast
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.project_1.R
+import com.example.project_1.ViewUserDetailFragment
 import com.example.project_1.data.remote.showPost.ShowPostData
 import com.example.project_1.data.remote.showPost.ShowPostRetrofitObject
+import com.example.project_1.ui.MainActivity
 import com.example.project_1.ui.showPost.adapter.PostAdapter
 import retrofit2.Call
 import retrofit2.Callback
@@ -20,43 +24,37 @@ class ShowPostFragment : Fragment() {
 
     private lateinit var postAdapter : PostAdapter
     private lateinit var rvShowPost : RecyclerView
+    private lateinit var showPostViewModel : ShowPostViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         val view = inflater.inflate(R.layout.item_post, container, false)
+        val homeButtonItemPost: Button = view.findViewById(R.id.btnHome)
 
         rvShowPost = view.findViewById(R.id.recyclerViewLoadImage)
         rvShowPost.layoutManager = LinearLayoutManager(requireContext())
         postAdapter = PostAdapter(emptyList())
         rvShowPost.adapter= postAdapter
-
+        showPostViewModel = ViewModelProvider(this).get(ShowPostViewModel::class.java)
         fetchPost()
 
+        homeButtonItemPost.setOnClickListener{
+           val mainActivity = activity as? MainActivity
+            mainActivity?.loadFragment(ViewUserDetailFragment())
+        }
         return view
     }
 
     private fun fetchPost() {
-        val inst = ShowPostRetrofitObject().getShowPostRetrofitObject()
-        inst.getPost().enqueue(object : Callback<List<ShowPostData>>{
-            override fun onResponse(
-                call: Call<List<ShowPostData>>,
-                response: Response<List<ShowPostData>>
-            ) {
-                if(response.isSuccessful){
-                    response.body()?.let { ShowPostData ->
-                        postAdapter.updateData(ShowPostData)
-
-                    }
-                }
-            }
-
-            override fun onFailure(call: Call<List<ShowPostData>>, t: Throwable) {
-                Toast.makeText(requireContext(), "Failed to fetch data", Toast.LENGTH_LONG).show()
+        showPostViewModel.fetchPosts(
+            onResult= { showPostData ->
+                postAdapter.updateData(showPostData)
 
             }
-        })
+        )
     }
+
 
 }
